remove next from delete unverified users fn -- move fn from middleware to lib folder
