+++ lib/delete-unverified-users.js
@@ -1,19 +1,15 @@
 import User from "../models/user-model.js";
 
 /**
- * @typedef {Function} ExpressNextFunction
- *
  * @async
  * @function
- * Middleware to be run as cron job to delete unverified users.
+ * To be run as cron job to delete unverified users.
  *
  * (See {@link server.js})
  *
- * @param {ExpressNextFunction} next - Function to call the next middleware.
- *
  * @returns {void}
  */
-const deleteUnveriedUsers = async (next) => {
+const deleteUnveriedUsers = async () => {
   try {
     const deletedUsers = await User.deleteMany({ verified: false });
 
@@ -25,7 +21,6 @@ const deleteUnveriedUsers = async (next) => {
   } catch (err) {
     console.log("Error deleting unverified users:", err);
   }
-  next();
 };
 
 export { deleteUnveriedUsers };
